+++ document-store/src/main/java/org/hypertrace/core/documentstore/mongo/query/parser/MongoIdentifierExpressionParser.java
@@ -1,7 +1,9 @@
 package org.hypertrace.core.documentstore.mongo.query.parser;
 
 import lombok.NoArgsConstructor;
+import org.hypertrace.core.documentstore.expression.impl.AliasedIdentifierExpression;
 import org.hypertrace.core.documentstore.expression.impl.IdentifierExpression;
+import org.hypertrace.core.documentstore.mongo.MongoUtils;
 
 @NoArgsConstructor
 public final class MongoIdentifierExpressionParser extends MongoSelectTypeExpressionParser {
@@ -16,7 +18,17 @@ public final class MongoIdentifierExpressionParser extends MongoSelectTypeExpres
     return parse(expression);
   }
 
+  @SuppressWarnings("unchecked")
+  @Override
+  public String visit(final AliasedIdentifierExpression expression) {
+    return MongoUtils.PREFIX + parse(expression);
+  }
+
   String parse(final IdentifierExpression expression) {
+    return expression.getName();
+  }
+
+  String parse(final AliasedIdentifierExpression expression) {
     return expression.getName();
   }
 }
