+++ document-store/src/main/java/org/hypertrace/core/documentstore/mongo/query/parser/MongoLetClauseBuilder.java
@@ -1,6 +1,7 @@
 package org.hypertrace.core.documentstore.mongo.query.parser;
 
-import com.mongodb.BasicDBObject;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.Map;
 import org.hypertrace.core.documentstore.expression.impl.AggregateExpression;
 import org.hypertrace.core.documentstore.expression.impl.AliasedIdentifierExpression;
@@ -13,6 +14,7 @@ import org.hypertrace.core.documentstore.expression.impl.KeyExpression;
 import org.hypertrace.core.documentstore.expression.impl.LogicalExpression;
 import org.hypertrace.core.documentstore.expression.impl.RelationalExpression;
 import org.hypertrace.core.documentstore.expression.type.FilterTypeExpression;
+import org.hypertrace.core.documentstore.mongo.MongoUtils;
 import org.hypertrace.core.documentstore.parser.FilterTypeExpressionVisitor;
 import org.hypertrace.core.documentstore.parser.SelectTypeExpressionVisitor;
 
@@ -26,79 +28,75 @@ class MongoLetClauseBuilder implements FilterTypeExpressionVisitor {
   }
 
   @Override
-  public BasicDBObject visit(LogicalExpression expression) {
-    BasicDBObject letClause = new BasicDBObject();
+  public Map<String, Object> visit(LogicalExpression expression) {
+    Map<String, Object> letClause = new HashMap<>();
     for (FilterTypeExpression operand : expression.getOperands()) {
-      letClause.putAll((Map<String, Object>) operand.accept(this));
+      letClause.putAll(operand.accept(this));
     }
-    return letClause;
+    return Collections.unmodifiableMap(letClause);
   }
 
   @Override
-  public BasicDBObject visit(RelationalExpression expression) {
-    BasicDBObject letClause = new BasicDBObject();
-    letClause.putAll(
-        (Map<String, Object>)
-            expression.getLhs().accept(new MongoLetClauseSelectTypeExpressionVisitor()));
-    letClause.putAll(
-        (Map<String, Object>)
-            expression.getRhs().accept(new MongoLetClauseSelectTypeExpressionVisitor()));
-    return letClause;
+  public Map<String, Object> visit(RelationalExpression expression) {
+    Map<String, Object> letClause = new HashMap<>();
+    letClause.putAll(expression.getLhs().accept(new MongoLetClauseSelectTypeExpressionVisitor()));
+    letClause.putAll(expression.getRhs().accept(new MongoLetClauseSelectTypeExpressionVisitor()));
+    return Collections.unmodifiableMap(letClause);
   }
 
   private class MongoLetClauseSelectTypeExpressionVisitor implements SelectTypeExpressionVisitor {
     @Override
-    public BasicDBObject visit(AggregateExpression expression) {
-      return new BasicDBObject();
+    public Map<String, Object> visit(AggregateExpression expression) {
+      return Collections.emptyMap();
     }
 
     @Override
-    public BasicDBObject visit(ConstantExpression expression) {
-      return new BasicDBObject();
+    public Map<String, Object> visit(ConstantExpression expression) {
+      return Collections.emptyMap();
     }
 
     @Override
-    public BasicDBObject visit(ConstantExpression.DocumentConstantExpression expression) {
-      return new BasicDBObject();
+    public Map<String, Object> visit(ConstantExpression.DocumentConstantExpression expression) {
+      return Collections.emptyMap();
     }
 
     @Override
-    public BasicDBObject visit(FunctionExpression expression) {
-      return new BasicDBObject();
+    public Map<String, Object> visit(FunctionExpression expression) {
+      return Collections.emptyMap();
     }
 
     @Override
-    public BasicDBObject visit(IdentifierExpression expression) {
-      return new BasicDBObject();
+    public Map<String, Object> visit(IdentifierExpression expression) {
+      return Collections.emptyMap();
     }
 
     @Override
-    public BasicDBObject visit(AliasedIdentifierExpression expression) {
+    public Map<String, Object> visit(AliasedIdentifierExpression expression) {
       return createLetClause(expression, subQueryAlias);
     }
-  }
 
-  @Override
-  public BasicDBObject visit(KeyExpression expression) {
-    return new BasicDBObject();
-  }
-
-  @Override
-  public BasicDBObject visit(ArrayRelationalFilterExpression expression) {
-    return new BasicDBObject();
-  }
-
-  @Override
-  public BasicDBObject visit(DocumentArrayFilterExpression expression) {
-    return new BasicDBObject();
-  }
-
-  private BasicDBObject createLetClause(
-      AliasedIdentifierExpression aliasedExpression, String subQueryAlias) {
-    BasicDBObject letClause = new BasicDBObject();
-    if (aliasedExpression.getContextAlias().equals(subQueryAlias)) {
-      letClause.put(aliasedExpression.getName(), "$" + aliasedExpression.getName());
+    private Map<String, Object> createLetClause(
+        AliasedIdentifierExpression aliasedExpression, String subQueryAlias) {
+      Map<String, Object> letClause = new HashMap<>();
+      if (aliasedExpression.getContextAlias().equals(subQueryAlias)) {
+        letClause.put(aliasedExpression.getName(), MongoUtils.PREFIX + aliasedExpression.getName());
+      }
+      return Collections.unmodifiableMap(letClause);
     }
-    return letClause;
+  }
+
+  @Override
+  public Map<String, Object> visit(KeyExpression expression) {
+    return Collections.emptyMap();
+  }
+
+  @Override
+  public Map<String, Object> visit(ArrayRelationalFilterExpression expression) {
+    return Collections.emptyMap();
+  }
+
+  @Override
+  public Map<String, Object> visit(DocumentArrayFilterExpression expression) {
+    return Collections.emptyMap();
   }
 }
